Rewrite buff tracker to understand worlds and be easier to configure.
The original implementation mixed up the code and buff definitions,
which made it hard to understand. It also had the unfortunate habit of
/echo'ing messages to the foreground world instead of the world which
triggered the message.

The rewrite scopes buffs to the world they're defined on, and also
provides some helper macros to quickly define when a buff starts and
ends. As a bonus, all the helper triggers are now invisible by default,
which makes /listdef much cleaner.