Improve buff tracker to allow config reloads with duplicating triggers
Now tracks the macro IDs for each trigger so they can be reset, allowing
for config file reloads. Before this, reloading a config file would
cause all the unnamed triggers for buffs to be redefined. Now, a config
file can invoke /reset_buffs <world> at the start to unset all variables
and triggers for tracking buffs in that particular world.